Update the Trees while still in the editor
So, our tree sprites change to match the tree type and the light
level. While this worked in game, the editor showed the trees as being
the default Light Elm. This made it hard to guess exactly what the
scene would look like until we ran the project.

Using the "tool" keyword and some setters (which I didn't know
existed!), the Tree scene will now update it's sprite AND collision
mode - IN THE EDITOR - when the user sets the export values. Neat!

Also, while I was at it, I updated the tree's collision masks. They
had been marked as "terrain" when they were supposed to be marked as
"scenery (blocking)".