Don't use mipmaps for Trees
The tree sprites had mipmaps enabled, but the mipmaps looked awful.
The mipmapped trees were small, thin, and wiry - so the trees looked
to be shrinking and growing.

I suspect this has something to do with how small the pixel trees are.
Because they're so small, maybe the mip-mapped texture is coming out
as mostly alpha pixels?

Doesn't really matter now - in this instance, I felt the trees looked
better without mipmaps, so they have been re-imported with mipmaps
disabled.