Add a third sub-grid to the valve grid
The Units align on a grid that's spaced out every .25 space-units. The
valve-style grid has a main grid (squares sized 1 space-unit) and a
sub grid (squares sized 0.5 space-units).

However, as we develop terrain and obstacles, the .25 space-units will
become the basis for how large these terrain/obstacle props are.

So, to ensure we're making these props to size, I've adjusted our
three grid textures. I added dots to the center of the sub-grids,
which denotes the .25 space-units. Thus, the three tiers of
demarcation (grid, sub-grid, dot) allow us to measure out three tiers
of space-units: .25, .50, and .75.