Begin work on the next terrain component: the Farm
The trees were our first piece of terrain. I'm happy with them - FOR
NOW!

The next terrain type I want to add is a farm. Farms may slow down the
pawns as they step through the crops, so a good pathing algorithm
would bypass it - unless the Pawn's goal is immediately inside the
farm or adjacent.

Before we can do anything crazy like pathing, we have to make the
farm. Each farm will be a variously-sized patch of dirt with some kind
of crop on it. We will most likely need to come up with various sizes,
shapes, and layout configurations.

For now, we have a same-sized dirt patch with some transition grass on
it. The dirt sits juuuuust above the grass (0.001 units, to be exact).
In the orthograpgic camera, with the grass plainly colored, it's
impossible to notice.

I was experimenting with having a plain dirt layer that was covered
with an alpha'd grass transition. The grass had to hover at 0.002
units, which is apparently too much because the effect was very
obvious. I may, at some point, have to try and circumvent Godot's
positional sig-figs being limited to 3. Perhaps with large scenes that
I then scale down?

For now, we'll need to come up with some sort of modular system (or
an alternative).