Add more sizes for the farm (and use dithering)
The initial texture for the farm plot used SURT's grass-to-dirt
transitions. They looked pretty good, but I wanted to try something
else.

So I made a texture that used dithering to slowly transition out of
the dirt pile. The effect is closer to the-messy-patch-of-dirt look I
was going for, but is that necessarily better than the SURT
transitions? I'm not sure.

One of the biggest questions right now is how apparent/scaled up the
pixels should be. In other words, what's the ideal pixel density? How
many pixels per unit of space? I want the pixelated aesthetic, I just
need to find the right size. If I can find a size I like, and stick
with it, the textures will be a lot more consistent.

I think that right now, with the farm, is probably the ideal time to
try out different pixel densities, with various types of crops.

As part of this, I want farms of dynamic size, with the crops
expanding and shrinking to match the field size. However, until I
select an appropriate pixel density, I can't really work on a tile
series or a tiling algorithm.

So, for now, I added a series of four plots of different sizes, all
using the dither texture. I added code to the farm to easily switch
between the plots.

Now with the plots semi-stable, I can work on making crops...