Scale up everything (starting with the UnitPawns)
As I expressed in previous commits, we were running into problems
involving our current scale. The UnitPawns were extremely small in
scale, and everything was scaled around that.

However, the scale and position vectors are limited to four digits.
We started to really run into issues with precision, as our core unit
was .1 - every other item was scaled and measured off of .1 units. So
half that was already .0125; we we already out of decimal places! We
couldn't quarter it, or eighth it, or anything like that.

I was trying to find a solution online - apparently the configurable
precision is an ongoing discussion/fix in Godot. But, to be honest, I
don't think that's a workable solution. There's only on true solution:
scaling everything up so we have more numbers to work with.

We're starting with the UnitPawn. I scaled up the UnitPawn, which made
the other UnitPawn variants out-of-date. I deleted those rather than
updating them, since we weren't using them anyway.

The new and improved UnitPawns are now 2 meters tall. Right now, only
the UnitPawns have been updated, so we got giant clowns on a small
board.