Scale up the trees 20x
Scaled up the tree scene by 20, to bring it in line with the rest of
the scaling changes.

All the sample trees had been pruned from ParadeGround during the
initial scaling procedure. This commit adds them back.

With the larger scale trees, I was able to more precisely align the
sprites with "Ground Level", so we sorta get the effect of roots on
the trees. The scaling is already paying dividends!

I need a better way to place trees - to put that another way, I need a
more deliberate way to place the trees. Right now I'm hand placing
them, but I think I'll want an algorithmic way to place them at some
point. That feature will probably be bundled as some sort of terrain
randomizer.