Correct shadows for sprites
Once I scaled up the UnitPawn sprites, the shadows were all wacky.
They had been configured for vastly smaller scales and caused weird
errors at the new scale, especially this weird shade-banding
phenomenon. The DirectionLight in ParadeGround was deleted and then
recreated as part of the scaling process, which fixed that issue.

However, looking at the UnitPawn sprites, I noticed the shadows
weren't connecting to the Pawn's feet. It looked like Peter Pan. We've
already done this before at small scale - now we need to do it at the
large scale. And this time I'll record EXACTLY WHAT I DID.

I immediately messed with the DirectionalLight's DirectionalShadow
property. I was previously able to avoid sprite-shadow cut off by
lowering the "Normal Bias" value. I lowered it from the default 0.8 to
0.1. I could now see the Pawn's whole shadow, but it wasn't connecting
at the feet.

To help with that, I looked at the node's "Light" properties. By
increasing the "Contact" value from 0 to 0.4, I observed a small
shadow almost being "painted" in to connect the shadow to the feet.
I'm not sure how that algorithm works, but it works like a charm!