Adjust UnitPawn to reflect upscaled realities
I previously upscaled the UnitPawns, but there were some more changes
that needed to be made.

First, for the clowns, the Clown sprite just didn't fit in the
collision box. Or, rather, it fit too well, in that it was just
floating in the middle of the collision cylinder. Looking at the
sprites in-game, they were floating. It was obnoxious!

To address this, I added two new debug animations to the Clown: the
sprite that appeared to be the "lowest" and the sprite that appeared
to be the "highest". This helped to establish an upper and lower limit
on how tall the clowns should be - the ideal height ended up being
1.75 units.

Secondly, the clowns were moving slow. Incredibly slow. Because they
were moving at their original speed for the 1/20th scale. I scaled up
the speed from [0.5 - 1.5] to [1 - 10].

Finally, I changed how we calculate the UnitPawn's "base" for the
purpose of goal checking. HEIGHT is gone; it's been replaced by
FLOOR_DISTANCE and an explanation for the where/why ethos behind the
values.