Update how the UnitPawns move
Previously, we were moving the UnitPawns with some sort of weird
misunderstanding of the relationship between vectors, distances, and
directions. I was using the total distance to the target as the
velocity, and... it was weird.

We now use a more proper algorithm - we even use the UNIT CIRCLE!

The UnitPawn's current speed/velocity is now set using linear values -
like a train moving forward and backward on a track. Separately, we
calculate the angle we need to move given our target on X and Z. This
is then fed into values using cos() and sin(). This gives us a
directional velocity, which we then scale with the line velocity.

We also now have a proper acceleration constant - neat! I am
experiencing an issue though - the UnitPawns accelerate but come to a
halt instantly. Sometimes they're moving so fast they are stuck in perpetual
target overshoot. I'll need to create an algorithm that accelerates and
decelerates as appropriate.

For now, though, I'm going back to asset creation.