Add wheat (dear god)
Okay, so I wanted to add some dynamic wheat that grew or shrank with
the farm fields.

At first I thought - "that's easy! I'll just slap something together
with quads and and a plane!" The issue there was it didn't expand
easily - modifying meshes directly in the code was a headache and
inconsistent. Plus, with how Godot handles materials, we would have
needed to rescale the material's UV each time. This would have
affected all the wheat textures, globally, which would have been a
giant mess.

I also had the idea for the field to "part" as creatures / NPCs walked
through it. The mesh wasn't working, so I was about to shelve that
whole plan, when I had an idea - what if we used Godot's 3D GridMap to
place individual "units" of wheat that could be added and removed
dynamically.

This turned out to be a terrible idea. Not because it didn't work, but
purely because of the art pipeline.

You see, the GridMap requires a resource called a "MeshLibrary". This
is a scene of nothing but meshes that gets converted into a special
format for use with the GridMap. The only issue was that I was making
my "Wheat Cubes" out of three separate meshes instead of one cohesive
mesh. This was translating in the MeshLibrary as a single
upward-facing plane.

So I needed to make a Cube capable of using multiple materials. And
that involved me learning how to use Blender. I was not happy, but I
did it.

I also had to make a texture for the top of the wheat (which I was
happy with) and the side of the wheat (which I am NOT HAPPY WITH). So
I had to put in way more effort than I wanted to and I hate the wheat
vomit texture I made but WE HAVE WHEAT!