Dynamically add size wheat on the the farm
So, we added wheat as a crop; now we need to dynamically expand the
size of the wheat.

The code to do this was pretty easy - open and shut really.

However, I'm starting to see limitations with the GridMap - there's no
real way to have each tile on the GridMap react dynamically. We cannot
attach things like lights or areas, or sprites. What we really need is
a NodeMap - which doesn't exist.

Maybe I should swap it - the farm plot should be built with the
GridMap, where the transition squares are part of the MeshLibrary.
Yeah, I think that would work...