Use GridMaps to create variable farm plots
We were using a series of textured planes to represent the different
sizes possible for each farm. However, this restricted what we could
do with the farm plots. Wouldn't it be better if we could size things
dynamically?

This commit adds a Mesh Library for the plot components. There are 9
total tiles - 4 edges, 4 coreners, and a center. The edge textures are
extra long and set to World Triplanar mode. This creates patterns that
are automatically different from one another and contigious. Neat!

My biggest concern right now is performance impacts. Just tiling stuff
by making a whole buncha meshes like that doesn't exactly sound like
the best idea, does it? I'm not seeing any noticeable or obvious
performance impacts yet, but I'll keep an eye out.