Use Nodes for Crops instead of GridMaps
Previously, we were using GridMaps for crops. This had the advantage
of speed and ease, but we couldn't interact with the individual grid
tiles. This was problematic - I had ideas about the wheat animating
when entered, or bursting into flames and slowly whittling away.
I couldn't see an easy way to do that on a per-tile basis.

The solution, then, was to place individual nodes, where each node had
whatever it needed to keep track of itself. I've removed the CropsGrid
from the Farm scene and replaced it with a generic Crops node. The
Wheat is now it's own scene, complete with some basic code to animate
whenever entered. It looks okay, not crazy about it, but I think
that's just fine.

I am seeing a performance impact, at least at the start. I'm going to
test on a lower-end system, just to ensure we're not losing
playability on potato machines. The isometric perspective cuts down on
our rendering burden, but we might be pushing it.