Experiment with Sprite crops
I had that block of wheat as my only crop - and I wasn't too happy
with it. I was worried about how spawning in all of those meshes could
have some bad performance impacts. Also, it didn't look great - though
that's something to be said about everything in this project. Still,
my roomate thought the wheat looked like cheese.

I was originally avoiding using sprites, since I thought that flat
sprites wouldn't be able to fill the screen and provide the density
that I would want/need them to. However - what would be the harm in
trying, if I was so unhappy with the cube-wheat?

So, I added a barley sprite, a node to match, and an option for the
farm to display that barley ("barley" being a natural cousin to
wheat). I added the barley sprite using a template for taller crops
and crops with underground portions - I'm not sure how crazy I am
about THAT just yet.

The node itself, however, looks much better than I thought it would.
I was initially using a "Sprite" for the barley node but saw a
considerable performance impact, so I switched to a Quad.

I wanted to try and replicate the barley getting crushed, so I put
some physics objects in it. While the effect doesn't look bad - I
might even call it good - the physics causes a massive drop in frames
and the barley freaks the hell out. I'll definitely use a different
technique, but I wanted this jank captured in a commit.