Add in a better "folding" system for Barley
After creating the barley and very dumb, physics based way to have it
"fold" when marched over, I need to make an improvement.

The Barley now uses an Area node to monitor/detect when a body
enters. When an appropriate body enters, the barley folds to lie flat
on the ground. As an added bonus, if the node has our
"_combined_velocity" variable (such as a UnitPawn), we fold it in the
direction that the object is moving. It actually looks really neat!

Of course, we need to keep an eye on performance. While I saw
considerable improvement over the wheat-cubes initially (frames rested
at ~100 fps), tracking each individual barley incurred a pretty hefty
toll (it now rests around ~70 fps). We'll keep an eye on the
performance impacts - we might have to skip on the folding if it's too
much of a burden.