Switch to sprite/quad based crops
After testing out how wheat/barley looked, I was really happy with
using 2D sprites for crops.

So, to follow up, I removed wheat from the official roster of crops
and added some new crops - cabbage, corn, carrots, and sunflowers.
Each one has it's own scene, so that we can define the behaviors
separately. For example, cabbage and carrots do not have the same
folding behavior that other crops do.

The effect looks really good. The art I made... It's awful. It's truly
awful. This is the first piece of programmer art that I'd classify as
adbominable. If I hire an artist to fix ANYTHING in this game, it will
be these crops.