Update how we "trample" crops
During tests, I noticed that there was some sort of weird effect on
the Barley when it was getting trampled. It looks like they were
casting shadows, which none of the other crops 'trampling' were doing.

At first, I thought it was the interplay between the clown shadows and
the barley stems, which normally blended into the dirt since they were
the same color and had no texture. I changed the barley to be brighter
but saw no improvement.

Then I figured I would just turn off the shadows once the trampling
happened, and that would fix it. Besides, that's a good improvement to
make anyway. However, the weird effect was still happening. I had to
change the camera to zoom in more just to find out what was happening:
for whatever reason, the barley wasn't laying flat on the ground. It
was hovering up in the air, meeting the clowns at their waists.

I'm still not exactly sure WHY, but the problem sprang from the order
of operations - the key sequence of rotations that made the crops
appear "trampled". I think it's because I was rotating X, then
rotating Y without restoring it to a 0-degree rotation. Somehow that
mishmash was creating floating barley, but just for barley. I'm not
sure, but I standardized my solution across all of the "trampling"
crops.

The solution was as such: rotate back the original angle before doing
further rotations, use radians instead of degrees, and using the
rotate_XYZ methods. That seems to have fixed it.

However, I was also happy with a lot of the changes I made trying to
fix the little issue I was pursuing - so I'm keeping those.