Add a house terrain piece
So I have trees, rocks, and crops with farms. For my final terrain
piece, I wanted to make a house, since buildings will feature
prominently in the next phase of this prototype.

Of course, given my Blender skills, I wanted to try and avoid working
in Blender AS MUCH AS POSSIBLE. My solution, then, was to approach the
house in the style of minecraft: I'd create a MeshLibrary of
components, and manually build up the house using a 3D GridMap.

The result is... definitely a house. To be honest, the use of a
GridMap harkens back to Space Engineers, or Minecraft. It doesn't look
as good as either of those things. However, it is functional, which is
good enough for me.