Add a to-do list
While I was working on this project yesterday, I found myself
paralyzed by indecision. I was thinking of like 50 different things to
do and the sum effect was I was doing none of them.

I need a clear(er) roadmap forward to help me focus on what needs to
be done. However, I didn't specifically want a 'full roadmap' because
I know my priorities and ideas are constantly shifting. What I really
need is a short term roadmap - a to-do list!

So, I added a to-do list. It's got everything I want done for the
immediate future on it. Neat!