Add a stupid death message scene
Okay, so in the last commit I laid out a gameplan for how I want to
address features in the immediate near term.

I immediately ignored this. I was listening to music and got inspired
to make dumb 3D text that changes color and bounces in time with some
music. It was a very dumb idea but I couldn't let it go and I really
wanted to try it. My idea was that, when the player lost because all
their units died, the music would play and the 3D 'you lost' message
would appear and bounce around with the music. And no, dying isn't
implemented, my brain is just running WAY AHEAD.

I created a model for the stupid text, but I haven't added music yet.
or death. I did, however, have to learn how to use Godot's shadering
language. I just used the Visual component, which was a little rougher
than I thought since all the tutorials were for the coding
alternative; the two did not correspond 1:1 as much as I would've
hoped. The result is... eh. I think I might be able to do something
better with a world-triplanar gradient texture...