Continue Messing with the Death Message Scene (WHY)
I just couldn't stop messing with the death message scene. I don't
know why. It was just too hard to resist.

First, I updated the text, swapping my broken shader attempt for a
prebaked gradient texture, with world-triplanar mode enabled. The sum
effect sorta approximates what I wanted - it's as if there's pockets
of color that the text is offering us a window into. Like a rainbow
cake, or something.

Secondly, I made the text animate, which is what I wanted it to do to
start with. I used a Tween to animate a custom property, allowing our
values to change over time. I had to do some funky stuff - especially
since 3D rotation with angles is frowned upon in Godot. That meant
precalculating and caching the rotations in Basis objects, then
capturing those in Quaternions, and then 'slerping' (???) between
them. I really didn't see how that was explicitly better than angles
but I suppose that's why I haven't made a game engine yet.

Finally, I painstakingly made the text animate at a very specific
rate: 134 bpm ...or, animations per minute, I guess. See, it was music
that inspired this stupid death screen. I was listening to "Your
Secrets" by SUI UZI when the idea of the bouncing animated text as a
game over screen came to mind. Obviously I won't be including the file
here, but I only mention it because I want to a) give credit and
b) document how I found the BPM.

You see, Audacity has a metronome track generator - they call it a
"Rhythm Track". Originally I was trying to count the beats compared to
the song but that was driving me mad. In a fit, I realized that I
could just ZOOM IN and match the waveforms. If the rhythm track had
too long between points, then the BPM needed to be higher. If it was
too short, the BPM needed to be lower. And, by god, it worked!

With this dumb death message out of the way, I should be able to get
back on track - starting with those dang villager houses...