Add in another flavor of medieval hovel
While I liked the little hexagonal hovel, it's technically inefficient
spacewise - since it doesn't have a neatly rectangular floorplan,
building up towns with it might be pretty annoying. Irritating even.
Plus, I wanted to explore more kinds of house.

So, I created a new kind of house, inspired by the medieval styles of
the homes in Rothenburg, Germany. The result is a home a lot like our
first two houses: a rectangular hut with a high peaked roof. I'm
pretty happy with the shape. I used the same technique as last time,
exporting the foundation, walls, and roof separate from each other.

I designed this house with "lower" and "upper" walls. The idea is two
fold: we fade out the roof and the upper walls, so they need to be
separate. At the same time, the upper walls are meant to be a bright
or pastel plaster material, while the lower walls are meant to be
stone.

For future reference, I've included the OpenSCAD file for reference.
As we make adjustments to the model, I'll move those changes into the
OpenSCAD file.

I haven't textured the houses yet - I just wanted to get them logged
in a commit first.