Redo the Rothenburg house
So I was trying to work with getting the Rothenburg house textured
(yes, actually textured, not just a flat color!) but I noticed that it
wasn't working. The textures were only showing up when I made the
material world triplanar. I remember having a similar issue with the
stupid cubes that I made for the wheat.

Turns out I didn't create a UV Map for the models, which, as I
understand it, roughly corresponds to the concept of 'Texture
Coordinates' that I am MUCH more familiar with. Basically, since I
didn't have any texture coordinates, we weren't getting any texture.

So, I added UV coordinates to each of the house's many components. To
make my life easier, I also now export all the meshes in a .dae file,
rather than exporting the models individually. I then duplicate this
to create the RothenburgHut scene.

One advantage of the .dae method - it recreates all the materials that
I made in Blender here. Normally I'd consider that an inconvenience
and unecessary clutter BUT in this instance, I was going to do that
anyway!