Discover how bad the Rothenburg Hut is via Textures
I didn't like the Rothenburg Hut, so I put textures on it.

Now I like it even less!

This thing is just a mess. The UV stuff I did worked, but not
particularly well. The foundations don't match up with the lower
walls, the upper walls are tilted (somehow?!?!), the windows don't
look that good considering their possible performance impact, and the
interior walls matching the outside actually just looks BAD.

Despite looking terrible, the parts of the Rothenburg Hut that DO work
actually look pretty good when textured! This not a fault of the
textures - it's the model. This model's gonna need a third edition...