Add stairs to the procedual house
Though I initially wanted to work on the house's lower walls, I put
some thought in and realized we needed some stairs. We're still a ways
out from solving how the UnitPawns move and perform "step-ups", so we
need a way for them to get on the foundations and into the houses in
the meantime.

So, I added stairs, with a new slate of configurables for controlling
the look and feel of the stairs. You can control the number of stairs
along with the length. One odd thing to note: the collision model
clips through the edges of the steps. I decided to do that because a
collision model large enough to enclose the stair edges would exceed
the user's specified lengths. Besides: with the UnitPawns
existing isometrically, the CollisionModel actually looks spot on.

While I was working on the sides of the stairs, I noticed that
plugging in vertices counter-clockwise (starting from the
lower-right) was creating the triangles on what I would have
designated the "back". So, I reversed the order and specified them
Clockwise from the upper-left. That ended up working out just fine.

But, when I was starting to line up the sides of the stairs with walls
of the foundation, I noticed they weren't lining up. After messing
around I noticed that this was because just the StairSide texture was
upside down. So, I inverted it in the material, but then I noticed
that the FoundationWall was ALSO upside down!

How did that happen? Did I not notice?! After some messing around, I
found out that what I thought was the "front" of the FoundationSide
was the actually the back, and I was only seeing the back because the
cull mode had somehow been set to "Both"?!? GAAAAAA!

While I was able to use this new info to cull the other components of
the foundation, it's become very apparent to me that something is
deeply wrong with how I've been making vertical faces. My horizontals
are fine, though. Maybe I should make a method for creating faces -
that could clean up a lot of my code too!