Add a back to the house's foundations
So I was initially of the mind that I didn't have to create polygon
faces for the backs of the houses, because having an isometric forced
perspective means we should really NEVER see the backs of the houses.

However, I realized that light is blocked (and shadow only created)
when the light hits the FORWARD face of a polygon surface/face. So if
I didn't create a back for the walls and foundations, light would
shine through into the house!

Since we don't want that, I added a back wall to the foundation, and I
will add a back to the walls when those eventually get added.

Also, I finally implemented the long-intended feature where you can
just tell the stairs to not generate. A simple addition, and it works.