Make buildings manually with procedural parts
I was starting to work on the "lower" walls for the procedural
generated building when I realized I was doing it all wrong - I was
trying to make the hut in such a way that, by messing with the
editor-exposed values, the user could create several different kinds
of buildings.

This meant that the components of the house, which each had their own
script, were not self-contained. That bugged me on a best practices
level. However, I also realized that this concept was useful for
quickly prototyping but NOT for creating buildings that deviated from
the template. In fact, it was more restrictive than anything else,
since trying to make all buildings as derivatives of a single building
would limit our possibilities.

What would work better would be to create procedual parts -
foundations, walls, stairs, windows, roofs, etc. - and then manually
construct buildings by hand using these procedual parts. This would
also open the door to do THAT procedually later on as well; if we can
do it by hand, we can do it procedually!

So, the existing stairs and foundation branches have been moved into
their own scenes. Their associated scripts have been updated so that
they now control their own values. It was pretty much a direct
translation. For example, the stairs currently have "shift" values for
X and Z so that they align much more easily with the foundations.
Since the stairs are only intended to be used WITH the foundations,
I'll consider that acceptable for now, and will be using similar
conventions for other procedual components.