Use translation for procedural parts, not "shifts"
I directly translated the values from RothStairs when I made it an
independent scene. One odd artifact of this was the "shift" values for
X and Y. I justified it in that, since these were only meant to be
used with other procedural parts and only in programmatic contexts,
this little odd-ball distinction wouldn't hurt too bad.

But then I was working on the wall components and I added a Y-shift.
I took a hard look and saw that I had just recreated the translation
value again, for some reason.

So I removed the shift values. We'll just use the translation values,
it's just easier.