Add wall components
I've added stairs and a foundation - next comes the walls!

The walls have controls for thickness, height, and length. This makes
it easy to build a house in-editor to specification. Though these
values are currently capped, they could theoretically be a larger or
smaller values - if I removed the caps. I won't - FOR NOW!

The walls have an interior and exterior side, as well as different
textures for the "cutaways" (the sides and top of the walls). The
walls are intended to be used as non-continous walls - i.e. each
"wall" component being one WALL of a house. For example, if you were
to just glue several wall blocks together, none of the textures would
line up!

However, sometimes you NEED these non-contiguous blocks to have
textures that line up. Maybe you're doing a more complex shape, like a
doorframe. In that case, I've provided a texture shift control. This
allows the user to manually line up textures over x, y, and z.

Adding in such a shift forced me to rework the PolyGen library. The
axis-locked functions now have one primary function that accepts
specific UV values. There are now several wrapper functions that allow
the user to pass various alternate values, which we then derive the UV
values from.

To test out the walls, I added them to a couple of the prototype
buildings I've been working on. They're not completely textured (since
I'm lacking textures for the cutaways) but they look pretty good - I'm
very satisfied!