Remove the Blender Rothenburg hut
The procedual buildings are coming along, and I want to clean up all
the assets I have just lying around.

To start, I've removed the Rothenburg-esque house that I made using
Blender. The textures were all screwed up and, to be frank, I wasn't
planning on using it for anything. It was just digital clutter.