Update procedural components on material change
Previously, we only did in-editor updates of procedural components
whenever we updated some of the constraining factors - size, height,
thickness, etc.

But we never updated when we changed the materials. To actually apply
our materials in-editor, we had to arbitrarily change one of the
constraints to force an update. This made things particularly
annoying, especially when working with a component that was otherwise
done correctly.

I've updated the scripts, adding setters for the materials so that the
editor meshes update appropriately.