Add columns
We've created procedual walls, but the walls in this game aren't meant
to meet. They were specifically designed with the intention that there
would be a different component that bridge the texture gap between
multiple walls.

This commit adds that component - the column! The column is a square
(since the walls should all be of the same thickness) and has the same
texture on all four sides, as well as a top texture.

I added some columns to our work-in-progress buildings along with some
different textures. I think I might play around with the columns in
the future - it might be worth it to have opposite sides with opposite
materials. For now, though, I'm more than happy!