Update the stairs
There were a couple of things about the "stairs" procedural component
that bugged me.

First, the "Sides" of the stairs were grouped by material and mesh to
both of the z-parallel faces, but also the x-parallel face. Part of
the reason I've avoided texturing the stairs is because I couldn't
think of a texture the would look good on both faces.

The solution, then, was to make the x-parallel face it's own mesh as
well as material. This should help make our stairs easier to manage,
though I might have to mess with the textures later on.

Secondly, I made the decision to have the stairs not extend to their
full height for some reason. Honestly I can't remember why. I think
that I thought it was disingeous to have a step match up with the
foundations when that step wasn't real.

But it looked ugly, and made texturing more difficult. So I've added a
toggle to go between the "full top step" stairs and the previous style
of stairs. Just in case I want those old stairs.