Constrain components using minimums, not exports
I previously defined the limits on procedural components using rough &
arbitrary values. These were also defined using the "export" keyword.

This presented multiple issues - firstly, with the arbitrary values. I
purposefully picked values that I thought were "good enough" but I
didn't really base that on any analysis of what the true values should
be. In most cases, there was no real maximum values, just minimum
values!

That leads to the problems with setting limits in the "export"
statement. To specify a minimum, we needed to specify a maximum -
which most of the procedural components don't even need! This was
compounded by the interface that consistently round values to only one
decimal place. Worse still - these values were only capped in the
editor. If I were to try and set the values programatically,
everything could go haywire!

The solution, then, was to constrain the components using a series of
minimum constants and the setter functions. This works much better and
the minimum values make a whole lot more sense. In most cases, I just
deferred to using .01, because many of the values are cut in half and
I don't want more than 3 decimal points.