Update frame thickness for foundations on size change
The MAXIMUM frame thickness for the Foundation procedural component is
a dynamic value - it changes with the sizes of the foundation.
However, we were only updating the frame thickness when we update the
frame thickness. Which is obvious, but we have to change it when the
dependent values update.

This has been changed, so the thickness should update appropriately as
the size changes.