Add a wall with a gap
We've built a couple of houses and I've identfied the need for some
more complex wall components. To start, we need a wall with a gap in
it. We've been doing this with two individual wall components, but we
can tie the functionality together into a more neat package.

This new wall is called RothWallOpenGap - it comes complete with all
the configurable items you'd come to expect. It's called an "Open Gap"
wall because the gap is open at the top, rather than having a third
wall that covers the gap. We'll need to create a separate scene just
for a wall with a closed gap.

To test it out, the gap-ed walls have been replaced in the RothHouse.
It looks good and seems functional so I'm pretty happy.

The hope is that, eventually, we can create a scene that just creates
four walls, columns and all. This is just the first step!