Add a "CloseGap" wall
The last commit added a wall with a gap - effectively, an easy
shortcut to control the size, spacing and materials of a couple of
walls.

The gap was really meant to represent a doorway, and MOST doorways
have a cap - they are but a hole in a wall, not a gap. So, to make a
more complete wall, I've added a "Close Gap" wall.

It's more or less a duplicate of the "Open Gap" wall, just with an
extra wall and a couple of extra export variables to match. The code
is also largely a duplicate - I know it'd technically be more
efficient to have this close gap wall exist as a subscene of the open
gap but that'd save... what, 200 line of code? The savings would be
trivial and I don't see future adjustments to these walls.