Add a component to easily create 4-wall rooms
As I mentioned before, I wanted to limit the number of materials
shared across multiple scenes and nodes. It makes more sense to me
that you should go to the relevant scene to edit the materials, and I
don't want changes to materials in one scene to effect other scenes.

In addition, while making buildings out of the walls and columns was
easy, it was resulting in cluttered scene trees and there weren't a
whole lot of easily recognizable names for me to use. It was starting
to get confusing.

So, I added a new procedural component to address these need. It
creates a room with four columns, at least three walls, and an
optional fourth wall. This new component, dubbed the Wall Set, allows
us to easily create rooms of various sizes on the fly. We can control
the height, the thickness, the size, and all the materials that go
into the various subcomponents.

This has allowed us to replace our existing wall structures in both
the house and the guardpost. That, in turn, allowed me to delete the
material resources. You now have to edit them in scene - just like I
wanted!