Only use one gap wall; spawn the "cap" as needed
When I created the two gap walls, my ethos was that the extra effort
needed to create a singular wall wasn't worth the benefit. However,
after implementing the wall-spawning in the WallSet, I realized that
we could actually do this with little effort. It'd be more compact and
I wouldn't need two separate scenes.

So I deleted the open gap wall - since that had the least amount of
code work - and modified the closed gap wall. There is now a boolean
option to add the cap, if needed. The WallSet has been updated
appropriately.